+++ lib/Pages/Goal/editGoalsPage.dart
@@ -1,4 +1,4 @@
-import 'package:acoin/db_context.dart';
+import 'package:acoin/utils/db_context.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/rendering.dart';
