+++ lib/Pages/dashboard.dart
@@ -1,21 +1,21 @@
-import 'package:acoin/DebtPage.dart';
-import 'package:acoin/category.dart';
 import 'package:acoin/data/database_helper.dart';
-import 'package:acoin/expense.dart';
-import 'package:acoin/income.dart';
-import 'package:acoin/loginPage.dart';
-import 'package:acoin/recurrentIncome.dart';
-import 'package:acoin/addExpensePage.dart';
-import 'package:acoin/addIncomePage.dart';
-import 'package:acoin/expensesHistoryPage.dart';
-import 'package:acoin/incomeHistoryPage.dart';
-import 'package:acoin/recurrentIncomeHistoryPage.dart';
-import 'package:acoin/db_context.dart';
-import 'package:acoin/slide_left_transition.dart';
+import 'package:acoin/Models/expense.dart';
+import 'package:acoin/Models/income.dart';
+import 'package:acoin/Models/recurrentIncome.dart';
+import 'package:acoin/Models/category.dart';
+import 'package:acoin/Pages/Expense/addExpensePage.dart';
+import 'package:acoin/Pages/Expense/expensesHistoryPage.dart';
+import 'package:acoin/Pages/Income/addIncomePage.dart';
+import 'package:acoin/Pages/Income/incomeHistoryPage.dart';
+import 'package:acoin/Pages/Income/recurrentIncomeHistoryPage.dart';
+import 'package:acoin/Pages/Goal/goalsPage.dart';
+import 'package:acoin/Pages/Debt/DebtPage.dart';
+import 'package:acoin/Pages/Identity/loginPage.dart';
+import 'package:acoin/Utils/db_context.dart';
+import 'package:acoin/Utils/slide_left_transition.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:acoin/goalsPage.dart';
 import 'package:flutter_speed_dial/flutter_speed_dial.dart';
 import 'package:charts_flutter/flutter.dart' as charts;
 import 'package:flushbar/flushbar.dart';
