+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:acoin/splashScr.dart';
+import 'package:acoin/Pages/splashScr.dart';
 import 'package:flutter/material.dart';
 //import 'buildExpensesPage.dart';
